more organized...?

+++ make-appimage.gtk.sh
@@ -2,6 +2,8 @@
 
 set -eu
 
+EXTR="peazip-gtk"
+
 ARCH="$(uname -m)"
 VERSION="v$(sed -n 1p version.txt)"
 
@@ -14,7 +16,7 @@ export ARCH VERSION
 export OUTPATH=./dist
 #export ADD_HOOKS="self-updater.bg.hook"
 #export UPINFO="gh-releases-zsync|${GITHUB_REPOSITORY%/*}|${GITHUB_REPOSITORY#*/}|latest|*$ARCH.AppImage.zsync"
-export ICON="peazip-gtk/res/share/icons/peazip.png"
+export ICON="$EXTR""/res/share/icons/peazip.png"
 export OUTNAME="$APPIMAGE_STEM".AppImage
 export DESKTOP="peazip.desktop"
 
@@ -33,13 +35,21 @@ export DEPLOY_SDL=0
 export DEPLOY_GLYCIN=0
 export APPDIR=./appdir-gtk
 
-mkdir -p "$APPDIR"/bin
-cp -va peazip-gtk/* "$APPDIR"/bin/
+if ! [ -d "$EXTR" ]
+then
+	echo "not found: $EXTR"
+	exit 0
+fi
+
+mkdir -vp "$APPDIR"/bin
+cp -va "$EXTR"/* "$APPDIR"/bin/
 rm -vrf "$APPDIR"/bin/res/portable
 rm -vrf "$APPDIR"/bin/res/conf
 
 # Deploy dependencies
-./quick-sharun.sh ./peazip-gtk/res/* ./peazip-gtk/peazip ./peazip-gtk/pea
+
+# ./quick-sharun.sh ./"$EXTR"/res/* ./"$EXTR"/peazip ./"$EXTR"/pea
+./quick-sharun.sh "$APPDIR"/bin/res/* "$APPDIR"/bin/peazip "$APPDIR"/bin/pea
 
 # Additional changes can be done in between here
 
@@ -51,15 +61,25 @@ then
 fi
 
 # Copy details
+
 mkdir -v "$APPDIR"/_details
 echo "$UBID" > "$APPDIR"/_details/commit.txt
 echo "$(date)" > "$APPDIR"/_details/date.txt
 rpm -qa > "$APPDIR"/_details/packages.txt
-sed -n 3p version.txt > "$APPDIR"/_details/upstream.txt
+
 fastfetch|sed -e 's/Local IP.*//' -e 's/Locale.*//' -e 's/Battery.*//' -e 's/Disk.*//' -e 's/Swap.*//' > "$APPDIR"/_details/system.txt
 
+US_FILE=$(ls|awk "/peazip_portable/ && /GTK/ && /$VERSION/ && /$ARCH/")
+US="$APPDIR"/_details/upstream.txt
+touch "$USR"
+echo "
+url" >> "$US"
+awk "/peazip_portable/ && /GTK/ && /$VERSION/ && /$ARCH/" sources.txt >> "$US"
+echo "
+sha256" >> "$US"
+sha256sum "$US_FILE" >> "$US"
+
 # Copy Internal scripts
-# mkdir -vp "$APPDIR"/bin
 
 cp -v is_details "$APPDIR"/bin/details
 cp -v is_setup.1.sh "$APPDIR"/bin/setup
